# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MIU Export Hub API",
    "description": "Made in Uganda Export Hub \u2014 B2B export marketplace backend.\n\n**Authentication:** send JWTs in the `Authorization` header, not query params.\n- Protected routes: `Authorization: Bearer <access_token>`\n- `/auth/refresh` and `/auth/logout`: `Authorization: Bearer <refresh_token>`",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/buyer/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Buyer Register",
        "description": "Create a buyer account and send an email activation link.\nThe user must activate before signing in or accessing the dashboard.",
        "operationId": "buyer_register_api_v1_auth_buyer_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerRegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/buyer/activate": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Buyer Activate",
        "description": "Activate account from email link token; returns auth tokens for the buyer dashboard.",
        "operationId": "buyer_activate_api_v1_auth_buyer_activate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerActivateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/buyer/resend-activation": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Buyer Resend Activation",
        "description": "Resend activation email if the account exists and is not yet activated.",
        "operationId": "buyer_resend_activation_api_v1_auth_buyer_resend_activation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerResendActivationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/buyer/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Buyer Login",
        "operationId": "buyer_login_api_v1_auth_buyer_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/buyer/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Buyer Me",
        "operationId": "buyer_me_api_v1_auth_buyer_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/buyer/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Buyer Refresh",
        "operationId": "buyer_refresh_api_v1_auth_buyer_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerAuthResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerRefresh": []
          }
        ]
      }
    },
    "/api/v1/auth/buyer/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Buyer Logout",
        "operationId": "buyer_logout_api_v1_auth_buyer_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerRefresh": []
          }
        ]
      }
    },
    "/api/v1/auth/buyer/change-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Buyer Change Password",
        "operationId": "buyer_change_password_api_v1_auth_buyer_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/supplier/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Supplier Register",
        "operationId": "supplier_register_api_v1_auth_supplier_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/supplier/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Supplier Login",
        "operationId": "supplier_login_api_v1_auth_supplier_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/supplier/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Supplier Me",
        "operationId": "supplier_me_api_v1_auth_supplier_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierAccountSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/supplier/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Supplier Refresh",
        "operationId": "supplier_refresh_api_v1_auth_supplier_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierAuthResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerRefresh": []
          }
        ]
      }
    },
    "/api/v1/auth/supplier/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Supplier Logout",
        "operationId": "supplier_logout_api_v1_auth_supplier_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerRefresh": []
          }
        ]
      }
    },
    "/api/v1/auth/supplier/change-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Supplier Change Password",
        "operationId": "supplier_change_password_api_v1_auth_supplier_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/admin/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Admin Login",
        "operationId": "admin_login_api_v1_auth_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/admin/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Admin Me",
        "operationId": "admin_me_api_v1_auth_admin_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/admin/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Admin Refresh",
        "operationId": "admin_refresh_api_v1_auth_admin_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerRefresh": []
          }
        ]
      }
    },
    "/api/v1/auth/admin/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Admin Logout",
        "operationId": "admin_logout_api_v1_auth_admin_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerRefresh": []
          }
        ]
      }
    },
    "/api/v1/auth/admin/change-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Admin Change Password",
        "operationId": "admin_change_password_api_v1_auth_admin_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuyerActivateRequest": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 256,
            "minLength": 16,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "BuyerActivateRequest"
      },
      "BuyerAuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "account": {
            "$ref": "#/components/schemas/BuyerAccountSummary"
          },
          "onboarding_required": {
            "type": "boolean",
            "title": "Onboarding Required",
            "default": false
          },
          "activation_required": {
            "type": "boolean",
            "title": "Activation Required",
            "default": false
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "account"
        ],
        "title": "BuyerAuthResponse"
      },
...
      "BuyerRegisterRequest": {
        "properties": {
          "company": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Company"
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "password_confirm": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password Confirm"
          }
        },
        "type": "object",
        "required": [
          "company",
          "first_name",
          "last_name",
          "email",
          "password",
          "password_confirm"
        ],
        "title": "BuyerRegisterRequest",
        "description": "Buyer self-service sign-up with password."
      },
      "BuyerRegisterResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "activation_required": {
            "type": "boolean",
            "title": "Activation Required",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "message",
          "email"
        ],
        "title": "BuyerRegisterResponse"
      },
      "BuyerResendActivationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "BuyerResendActivationRequest"
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "type": "string",
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password"
          },
          "new_password_confirm": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password Confirm"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password",
          "new_password_confirm"
        ],
        "title": "ChangePasswordRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "SupplierAccountSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "must_change_password"
        ],
        "title": "SupplierAccountSummary"
      },
      "SupplierAuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "account": {
            "$ref": "#/components/schemas/SupplierAccountSummary"
          },
          "onboarding_required": {
            "type": "boolean",
            "title": "Onboarding Required",
            "default": false
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "account"
        ],
        "title": "SupplierAuthResponse"
      },
      "SupplierRegisterRequest": {
        "properties": {
          "company": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Company"
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "password_confirm": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password Confirm"
          }
        },
        "type": "object",
        "required": [
          "company",
          "first_name",
          "last_name",
          "email",
          "password",
          "password_confirm"
        ],
        "title": "SupplierRegisterRequest",
        "description": "Supplier self-service sign-up with password."
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "JWT access token. Example: Authorization: Bearer eyJhbGciOiJIUzI1NiIs...",
        "scheme": "bearer"
      },
      "BearerRefresh": {
        "type": "http",
        "description": "Refresh token for token rotation and logout.",
        "scheme": "bearer"
      }
    }
  }
}